requirements-test.txt
# Testing dependencies for PLM_SARS-CoV-2
# Install with: pip install -r requirements-test.txt

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-timeout>=2.1.0
pytest-mock>=3.10.0

# Code quality
flake8>=6.0.0
black>=23.0.0
pylint>=2.16.0
isort>=5.12.0

# Required for Functions_HuggingFace.py
biopython>=1.79
pandas>=1.3.0
numpy>=1.21.0
torch>=1.9.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
py3Dmol>=2.0.0
IPython>=7.0.0

# Optional but recommended
coverage[toml]>=7.0.0
pytest-html>=3.1.0
